09-full-stack/backend/requirements.txt
Flask
flask-sqlalchemy
flask-cors
python-dotenv
pytest
gunicorn